requirements.txt
flask
flask-cors
sqlalchemy
python-dotenv
groq
pyjwt
gunicorn
